requirements.txt
streamlit
python-chess
stockfish
pycairo
cairosvg



